chore: add typescript support
